agent/pyproject.toml
[project]
name = "agentforge"
version = "0.1.0"
description = "AI agent for Ghostfolio portfolio management"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "langgraph>=0.2",
    "openai>=1.0",
    "fastapi>=0.115",
    "uvicorn>=0.34",
    "httpx>=0.28",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "langsmith>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "respx>=0.22",
]

[build-system]
requires = ["setuptools>=75.0"]
build-backend = "setuptools.build_meta"
